fix(knowledge): address review feedback and fix noexec tmpdir test failures
- Extract duplicated `stub_popen3` helper from both collector specs into
  shared `spec/support/popen3_stub.rb` module, included via `type: :service`
- Reap child process after timeout kill in `BaseCollector#run_command` by
  calling `wait_thr.value` to prevent zombie process accumulation
- Fix dev_script and dev_update specs to use exec-allowed tmpdir when /tmp
  is mounted noexec
